Update to new Albert HTML look and feel
Around January 6, 2018, [Albert changed its look and feel](https://t.e2ma.net/message/riaqw/vtp8qz).  So the parser had to be modified.

The good news is that the file doesn't use frames anymore, so the parsing is streamlined a bit.  Also, there wasn't much that needed to be changed except the photo key and setting of the base directory.